+++ src/syscall/execve.rs
@@ -17,7 +17,7 @@ const SHEBANG_MAX: usize = 257;
 /// `file` is the file from which the shebang is to be read.
 /// `buff` is the buffer to write the shebang on.
 /// If the file has a shebang, the function returns its size in bytes.
-pub fn read_shebang(file: &File, buff: &mut [u8; SHEBANG_MAX]) -> Result<Option<usize>, Errno> {
+pub fn read_shebang(file: &File, buff: &mut [u8; SHEBANG_MAX]) -> Result<Option<u64>, Errno> {
 	let size = file.read(0, buff)?;
 
 	if size >= 2 && buff[0..1] == [b'#', b'!'] {
